+++ gazebo/gui/model/CollisionConfig.hh
@@ -145,6 +145,8 @@ namespace gazebo
 
       /// \brief Qt signal emitted when a collision is added.
       /// \param[in] _name Name of collision added.
+      /// \param[in] _collisionShape Collision shape, either "box", "sphere" or
+      /// "cylinder".
       Q_SIGNALS: void CollisionAdded(const std::string &_name,
         const std::string &_collisionShape);
 
@@ -157,6 +159,8 @@ namespace gazebo
       Q_SIGNALS: void Applied();
 
       /// \brief Qt callback when a collision is to be added.
+      /// \param[in] _collisionShape Collision shape, either "box", "sphere" or
+      /// "cylinder".
       private slots: void OnAddCollision(const QString &_collisionShape);
 
       /// \brief Qt callback when a collision is to be removed.
@@ -217,6 +221,7 @@ namespace gazebo
 
       /// \brief Qt signal mapper for mapping add collision button signals.
       private:  QSignalMapper *mapperAdd;
+
       /// \brief A map of collision items to their id.
       private: std::map<int, QTreeWidgetItem *> collisionItems;