+++ gazebo/gui/model/ModelData.hh
@@ -215,6 +215,8 @@ namespace gazebo
 
       /// \brief Qt callback when a new collision is to be added.
       /// \param[in] _name Name of collision.
+      /// \param[in] _collisionShape Collision shape, either "box", "sphere" or
+      /// "cylinder".
       private slots: void OnAddCollision(const std::string &_name,
         const std::string &_collisionShape);